chimera:adding labels metadata  to file objects
Currently  it is  not possible to create/add labels to file objects or perform a search based on labels.

The current patch aims to provide this functionalities for attaching labels to files and to give users a possibility to search for files with a given label.

Modifications

new t_label and t_label_ref tables are added to chimera.

`VirtualDirectoryStream(String labelname)` method is responsible for listing all files having `lablename` as a label metadata.

Acked-by:
Target: master
Require-book: yes
Require-notes: yes